rename random chords
