+++ chord-game/src/lib/RandomChords.svelte
